* commit 'c73b46f40a3e47ed9989cf1c9681901572b508be':
  Do not call exit 1 when loaded with source command

+++ tools/dicttool/tests/etc/test-dicttool.sh
@@ -18,7 +18,7 @@ echo "Usage:" 1>&2
 echo "    source $0" 1>&2
 echo "  or" 1>&2
 echo "    . $0" 1>&2
-exit 1
+if [[ ${BASH_SOURCE[0]} != $0 ]]; then return; else exit 1; fi
 fi
 
 find out -name "dicttool_aosp*" -exec rm -rf {} \; > /dev/null 2>&1
